+++ junit/kube-api-test/client-inject/src/test/java/io/fabric8/kubeapitest/junit/TestCaseUtils.java
@@ -14,48 +14,37 @@
  * limitations under the License.
  */
 
-package io.fabric8.kubeapitest.sample;
+package io.fabric8.kubeapitest.junit;
 
 import io.fabric8.kubernetes.api.model.ConfigMap;
 import io.fabric8.kubernetes.api.model.ConfigMapBuilder;
 import io.fabric8.kubernetes.api.model.ObjectMetaBuilder;
-import io.fabric8.kubernetes.client.Config;
 import io.fabric8.kubernetes.client.KubernetesClient;
-import io.fabric8.kubernetes.client.KubernetesClientBuilder;
 import org.assertj.core.api.Assertions;
 
 import java.util.Map;
 
-public class TestUtils {
+public class TestCaseUtils {
 
-  /**
-   * Used in multiple tests so no multiple additional binaries are downloaded, therefore increases
-   * duration of tests.
-   **/
-  public static final String NON_LATEST_API_SERVER_VERSION = "1.26.0";
+  public static final String TEST_1 = "test1";
 
-  public static ConfigMap testConfigMap() {
+  public static ConfigMap testConfigMap(String name) {
     return new ConfigMapBuilder()
         .withMetadata(new ObjectMetaBuilder()
-            .withName("test1")
+            .withName(name)
             .withNamespace("default")
             .build())
         .withData(Map.of("key", "data"))
         .build();
   }
 
-  public static void simpleTest(String kubeConfigYaml) {
-    simpleTest(
-        new KubernetesClientBuilder().withConfig(Config.fromKubeconfig(kubeConfigYaml)).build());
-  }
-
-  public static void simpleTest() {
-    simpleTest(new KubernetesClientBuilder().build());
-  }
-
   public static void simpleTest(KubernetesClient client) {
-    client.resource(TestUtils.testConfigMap()).create();
-    var cm = client.resource(TestUtils.testConfigMap()).get();
+    simpleTest(client, TEST_1);
+  }
+
+  public static void simpleTest(KubernetesClient client, String testResourceName) {
+    client.resource(TestCaseUtils.testConfigMap(testResourceName)).create();
+    var cm = client.resource(TestCaseUtils.testConfigMap(testResourceName)).get();
 
     Assertions.assertThat(cm).isNotNull();
