+++ junit/kube-api-test/client-inject/src/test/java/io/fabric8/kubeapitest/junit/sample/JUnitFabric8ClientInjectPerMethodTest.java
@@ -17,7 +17,7 @@
 package io.fabric8.kubeapitest.junit.sample;
 
 import io.fabric8.kubeapitest.junit.EnableKubeAPIServer;
-import io.fabric8.kubeapitest.junit.TestUtils;
+import io.fabric8.kubeapitest.junit.TestCaseUtils;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import org.junit.jupiter.api.Test;
 
@@ -28,12 +28,12 @@ class JUnitFabric8ClientInjectPerMethodTest {
   @Test
   @EnableKubeAPIServer
   void simpleTest1() {
-    TestUtils.simpleTest(client);
+    TestCaseUtils.simpleTest(client);
   }
 
   @Test
   @EnableKubeAPIServer
   void simpleTest2() {
-    TestUtils.simpleTest(client);
+    TestCaseUtils.simpleTest(client);
   }
 }
