+++ junit/kube-api-test/core/src/test/java/io/fabric8/kubeapitest/sample/JUnitExtensionKubeConfigUpdateTest.java
@@ -24,6 +24,6 @@ class JUnitExtensionKubeConfigUpdateTest {
 
   @Test
   void usesConfigFromGenericKubeConfig() {
-    TestUtils.simpleTest();
+    TestCaseUtils.simpleTest();
   }
 }
