+++ junit/kube-api-test/core/src/test/java/io/fabric8/kubeapitest/sample/TestCaseUtils.java
@@ -14,37 +14,48 @@
  * limitations under the License.
  */
 
-package io.fabric8.kubeapitest.junit;
+package io.fabric8.kubeapitest.sample;
 
 import io.fabric8.kubernetes.api.model.ConfigMap;
 import io.fabric8.kubernetes.api.model.ConfigMapBuilder;
 import io.fabric8.kubernetes.api.model.ObjectMetaBuilder;
+import io.fabric8.kubernetes.client.Config;
 import io.fabric8.kubernetes.client.KubernetesClient;
+import io.fabric8.kubernetes.client.KubernetesClientBuilder;
 import org.assertj.core.api.Assertions;
 
 import java.util.Map;
 
-public class TestUtils {
+public class TestCaseUtils {
 
-  public static final String TEST_1 = "test1";
+  /**
+   * Used in multiple tests so no multiple additional binaries are downloaded, therefore increases
+   * duration of tests.
+   **/
+  public static final String NON_LATEST_API_SERVER_VERSION = "1.26.0";
 
-  public static ConfigMap testConfigMap(String name) {
+  public static ConfigMap testConfigMap() {
     return new ConfigMapBuilder()
         .withMetadata(new ObjectMetaBuilder()
-            .withName(name)
+            .withName("test1")
             .withNamespace("default")
             .build())
         .withData(Map.of("key", "data"))
         .build();
   }
 
-  public static void simpleTest(KubernetesClient client) {
-    simpleTest(client, TEST_1);
+  public static void simpleTest(String kubeConfigYaml) {
+    simpleTest(
+        new KubernetesClientBuilder().withConfig(Config.fromKubeconfig(kubeConfigYaml)).build());
   }
 
-  public static void simpleTest(KubernetesClient client, String testResourceName) {
-    client.resource(TestUtils.testConfigMap(testResourceName)).create();
-    var cm = client.resource(TestUtils.testConfigMap(testResourceName)).get();
+  public static void simpleTest() {
+    simpleTest(new KubernetesClientBuilder().build());
+  }
+
+  public static void simpleTest(KubernetesClient client) {
+    client.resource(TestCaseUtils.testConfigMap()).create();
+    var cm = client.resource(TestCaseUtils.testConfigMap()).get();
 
     Assertions.assertThat(cm).isNotNull();
